fix/pseudowords: fix word_lexicon generation
Having `float(frequency_per_million) > cutoff` with `cutoff = 0` led to
excluding all words with frequency 0 from the lexicon. This showed up in
the fact that generating pseudowords for 'fell' included 'full' more
often than not, which is a real word. Setting the comparison to `>=`
fixes the problem.

Note that this never showed up for the original authors because they
forgot to convert frequency_per_million to a `float`, thus comparing a
`str` to an `int`, which in Python 2 always evaluates to `True`, making
that conditional effectively useless. In Python 3 it's an error, so
after fixing that bug, this new behaviour appeared.